+++ src/artistsmodel.cpp
@@ -72,7 +72,7 @@ namespace unplayer
 
     int ArtistsModel::rowCount(const QModelIndex&) const
     {
-        return mArtists.size();
+        return static_cast<int>(mArtists.size());
     }
 
     bool ArtistsModel::sortDescending() const
@@ -182,7 +182,7 @@ namespace unplayer
         }
         QObject::connect(LibraryUtils::instance(), &LibraryUtils::removingFilesChanged, this, [this, indexes] {
             if (!LibraryUtils::instance()->isRemovingFiles()) {
-                for (int i = indexes.size() - 1; i >= 0; --i) {
+                for (int i = static_cast<int>(indexes.size()) - 1; i >= 0; --i) {
                     const int index = indexes[i];
                     beginRemoveRows(QModelIndex(), index, index);
                     mArtists.erase(mArtists.begin() + index);
